# add a hack to avoid inference recursion limit to allow more optimization