--- a/Project.toml
+++ b/Project.toml
@@ -13,8 +13,9 @@ Parsers = "2.5.1"
 julia = "1.6"
 
 [extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test", "OrderedCollections"]
+test = ["Test", "JET", "OrderedCollections"]