@quinnj 2023-03-24
Woohoo! Thanks so much @aviatesk! Sorry for the slow response too; I've been swamped w/ some other things lately, so I've only had a little time to try and digest this and come up w/ a fix for https://github.com/quinnj/JSONBase.jl/issues/3 in [this PR](https://github.com/quinnj/JSONBase.jl/pull/5).

Given that we _can_ support structs w/ recursive fields, do you think there's any way we can still get the optimizations in this PR? I would be happy to somehow disable the optimizations just for _actually_ recursive structs if it meant getting the optimizations for all other structs. Or we could even go as far as requiring a specific overload for recursive structs so they take an entirely different codepath all together? Would love to hear what you think is possible here.