@aviatesk 2023-03-24
No worries, no hurry from my side at all :)

> Given that we _can_ support structs w/ recursive fields, do you think there's any way we can still get the optimizations in this PR?

I think we can still get the optimizations with supporting structs with recursive fields! I think this PR might work with #5 , but I'd like to check if we don't fail into the recursive inference. So could you merge #5 first? Then I will rebase this PR and do some experiments.