@quinnj 2023-03-24
Ok, I tried this PR out on latest master, but I'm seeing a slightly weird issue when running the original script:
```julia
using JSONBase

struct A
  a::Int
  b::Int
  c::Int
  d::Int
end

@btime JSONBase.materialize("""{ "a": 1,"b": 2,"c": 3,"d": 4}""", A)
```

(this is with latest Julia#master as of today, 2023-03-24).

At least once from a fresh REPL, it was _fast_ (the expected ~180ns), but the `pprof` flamegraph was weird w/ an entry for `ijl_box_int64` in the applyfield -> macro expansion spans.

Fast: 
<img width="1384" alt="image" src="https://user-images.githubusercontent.com/2896623/227638053-9f5db8fe-98f4-472e-8ee3-22bf4163c6c6.png">

But from another fresh REPL, I'm seeing it slower (around ~300ns), and the `pprof` flamegraph shows this `ijl_box_int64` and a `jl_f_fieldtype` that seem to relatively expensive to everything else going on.

Slow: 
<img width="1342" alt="image" src="https://user-images.githubusercontent.com/2896623/227638243-d7901002-ff03-4e18-90cf-d9d6dc8aa596.png">

I'll dig in a bit further, but it's unclear where the Int boxing is happening along w/ the `fieldtype` call?