@aviatesk 2023-03-27
Hmm, I couldn't reproduce the weird performance issue (it runs consistently fast). Do you have a step to reproduce the pprof graph?