@quinnj 2023-03-31
Yeah, so here's my versioninfo:

```julia
julia> versioninfo()
Julia Version 1.10.0-DEV.890
Commit 39fd7ddbee (2023-03-28 03:29 UTC)
Platform Info:
  OS: macOS (arm64-apple-darwin22.2.0)
  CPU: 10 × Apple M1 Max
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, apple-m1)
  Threads: 8 on 8 virtual cores
Environment:
  JULIA_PROJECT = @.
  JULIA_NUM_THREADS = auto
  JULIA_PKG_SERVER_REGISTRY_PREFERENCE = eager
```

and then I'm doing:

```julia
using JSONBase, BenchmarkTools, PProf
struct A
     a::Int
     b::Int
     c::Int
     d::Int
 end
@bprofile JSONBase.materialize("""{ "a": 1,"b": 2,"c": 3,"d": 4}""", A)
pprof()
```
Then click on the link in the pprof output which opens a local web browser page, then in the top left click "View" -> "Flamegraph".